+++ tests/Feature/ModularNumberTest.php
@@ -75,7 +75,7 @@ class ModularNumberTest extends TestCase
     #[DependsOnClass(DifferentModulusErrorTest::class)]
     #[DependsOnClass(OperationTest::class)]
     #[TestDox("can be added to another.")]
-    public function test_add_returns_modular_number(): void
+    public function test_addition(): void
     {
         // Arrange
         $a = $this->randomModularNumber(max: $this::MAX);
@@ -88,8 +88,22 @@ class ModularNumberTest extends TestCase
 
     #[DependsOnClass(DifferentModulusErrorTest::class)]
     #[DependsOnClass(OperationTest::class)]
+    #[TestDox("can be subtracted from another.")]
+    public function test_subtraction(): void
+    {
+        // Arrange
+        $a = $this->randomModularNumber(max: $this::MAX);
+        $b = $this->randomModularNumberWithModulus($a->modulus, max: $this::MAX);
+
+        // Act & Assert
+        $this->assertInstanceOf(ModularNumber::class, $result = $a->sub($b));
+        $this->assertEquals($a->value->sub($b->value)->mod($a->modulus)->value, $result->value);
+    }
+
+    #[DependsOnClass(DifferentModulusErrorTest::class)]
+    #[DependsOnClass(OperationTest::class)]
     #[TestDox("can be multiplied to another.")]
-    public function test_multiply_returns_modular_number(): void
+    public function test_multiplication(): void
     {
         // Arrange
         $a = $this->randomModularNumber(max: $this::MAX);
@@ -102,8 +116,24 @@ class ModularNumberTest extends TestCase
 
     #[DependsOnClass(DifferentModulusErrorTest::class)]
     #[DependsOnClass(OperationTest::class)]
+    #[TestDox("can be divide by another.")]
+    public function test_division(): void
+    {
+        // Arrange
+        $a = $this->randomModularNumber(max: $this::MAX);
+        do {
+            $b = $this->randomModularNumberWithModulus($a->modulus, max: $this::MAX);
+        } while ($b->value->isEqual(0));
+
+        // Act & Assert
+        $this->assertInstanceOf(ModularNumber::class, $quotient = $a->div($b));
+        $this->assertEquals($a->value->div($b->value)->mod($a->modulus)->value, $quotient->value);
+    }
+
+    #[DependsOnClass(DifferentModulusErrorTest::class)]
+    #[DependsOnClass(OperationTest::class)]
     #[TestDox("can be raised to power.")]
-    public function test_power_returns_modular_number(): void
+    public function test_power(): void
     {
         // Arrange
         $a = $this->randomModularNumber(max: 1000.0);